[python] Consolidate `test_util_tiledb.py` into `test_stats.py`
